+++ SIBI_Backend/SIBI_Backend/Modelos/Alquileres/EntradaAlquiler.cs
@@ -5,7 +5,7 @@
         public Guid idSocio {  get; set; }
         public DateTime fechaDesde { get; set; }
         public DateTime fechaHasta { get; set; }
-        public int montoTotal { get; set; }
+        public decimal montoTotal { get; set; }
         public int? puntosCanjeados { get; set; }
         public List<EntradaDetalleAlquiler> detalleAlquiler { get; set; }
     }
