+++ SIBI_Backend/SIBI_Backend/Modelos/Alquileres/EntradaDetalleAlquiler.cs
@@ -3,7 +3,7 @@
     public class EntradaDetalleAlquiler
     {
         public Guid idLibro {  get; set; }
-        public int precioAlquiler { get; set; }
-        public int subtotalDetalle {  get; set; }
+        public decimal precioAlquiler { get; set; }
+        public decimal subtotalDetalle {  get; set; }
     }
 }
